// libs/journeys/ui/src/components/RadioQuestion/__generated__/RadioOptionBlockCreate.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { RadioOptionBlockCreateInput } from "./../../../../__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: RadioOptionBlockCreate
// ====================================================

export interface RadioOptionBlockCreate_radioOptionBlockCreate {
  __typename: "RadioOptionBlock";
  id: string;
}

export interface RadioOptionBlockCreate {
  radioOptionBlockCreate: RadioOptionBlockCreate_radioOptionBlockCreate;
}

export interface RadioOptionBlockCreateVariables {
  input: RadioOptionBlockCreateInput;
}